+++ Push_swap/ft_printf/side_func.c
@@ -6,13 +6,13 @@
 /*   By: mdegache <mdegache@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/22 11:38:01 by mdegache          #+#    #+#             */
-/*   Updated: 2024/10/29 15:26:18 by mdegache         ###   ########.fr       */
+/*   Updated: 2024/11/28 09:11:49 by mdegache         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int	i;
